+++ docs/viewer/compare-composition-test.ts
@@ -1,11 +1,9 @@
-import { glob } from "glob";
 import { Ifc5FileJson } from "../../schema/out/@typespec/json-schema/ts/ifc5file";
 import { compose, ComposedObject } from "./compose";
 import { compose2 } from "./compose2";
 import { describe, each, it } from "./test/util/cappucino";
 import { expect } from "chai";
 import * as fs from "fs";
-import * as path from "path";
 
 function CompareComposition(a: ComposedObject, b: ComposedObject)
 {
@@ -77,13 +75,6 @@ function CompareComposition(a: ComposedObject, b: ComposedObject)
     return true;
 }
 
-const fixtureDirectories = glob.sync('test/fixtures/*');
-
-function Cleanup(obj)
-{
-    return JSON.parse(JSON.stringify((obj)));
-}
-
 describe("composition comparison", () => {
     it("should be equal for 'hello-wall.ifcx' and 'hello-wall_add-firerating.ifcx'", async() => {
         // arrange
@@ -102,20 +93,3 @@ describe("composition comparison", () => {
         expect(outcome).to.be.true;
     });
 });
-
-describe("composition", () => {
-    each("should properly handle fixture directory", fixtureDirectories, (fixtureDir) => {
-
-        const inputFiles = glob.sync(`${fixtureDir.replace(/\\/g, '/')}/input_*.ifcx.json`);
-        expect(inputFiles.length).to.be.above(0);
-        const inputs = inputFiles.map((inputFile) => {
-            return JSON.parse(fs.readFileSync(inputFile, 'utf8'));
-        });      
-        let actualResult = Cleanup(compose2(inputs as Ifc5FileJson[]));
-        const outputFile = path.join(fixtureDir, 'output.json');
-        const expectedOutput = JSON.parse(fs.readFileSync(outputFile).toString());
-        // console.log("actual", JSON.stringify(actualResult, null, 4));
-        // console.log("expected", JSON.stringify(expectedOutput, null, 4));
-        expect(actualResult).to.deep.equal(expectedOutput);
-    });
-})
