+++ docs/viewer/package.json
@@ -10,6 +10,7 @@
   "author": "",
   "description": "",
   "devDependencies": {
+    "@types/node": "^22.13.10",
     "@types/three": "^0.171.0",
     "chai": "^5.2.0",
     "esbuild": "^0.24.2",
