+++ docs/viewer/test.ts
@@ -1,8 +1,10 @@
 import { test } from "./test/util/cappucino";
 
 import * as CompareComposition from "./compare-composition-test"
+import * as CompositionTest from "./composition-test"
 
 
 CompareComposition
+CompositionTest
 
 test();
